Conflicts:
	scilab/Scilab.sln
	scilab/Scilab_f2c.sln
	scilab/configure
	scilab/configure.ac
	scilab/etc/modules.xml.in
	scilab/etc/modules.xml.vc
	scilab/modules/Makefile.am
	scilab/modules/Makefile.in
	scilab/modules/core/includes/callinterf.h
	scilab/modules/core/src/c/callinterf.c
	scilab/modules/core/src/c/dynamic_gateways.c
	scilab/modules/core/src/c/dynamic_gateways.h
	scilab/tools/innosetup/files.iss

+++ scilab/modules/graphics/src/nographics/nographics.c
@@ -180,7 +180,7 @@ sciPointObj *sciGetPointerFromHandle (long handle)
 /*--------------------------------------------------------------------------*/
 unsigned long getHandleFromStack( size_t stackPointer )
 {
-	return 	-1;
+	return 	(unsigned long)(-1);
 }
 /*--------------------------------------------------------------------------*/
 int callSetProperty( sciPointObj * pObj, size_t stackPointer, int valueType, int nbRow, int nbCol, char * propertyName )
