+++ scilab/modules/gui/sci_gateway/c/sci_uimenu.c
@@ -56,7 +56,7 @@ int sci_uimenu( char *fname,unsigned long fname_len )
     if (Rhs == 0)
     {
         // Set the parent property
-        setMenuParent(sciGetPointerFromHandle(GraphicHandle), -1, sci_handles, nbRow, nbCol);
+        setMenuParent(sciGetPointerFromHandle(GraphicHandle), (size_t)(-1), sci_handles, nbRow, nbCol);
     }
 
     /**
@@ -206,7 +206,7 @@ int sci_uimenu( char *fname,unsigned long fname_len )
     if (!parentDefined && (Rhs != 0))
     {
         // Set the parent property
-        setMenuParent(sciGetPointerFromHandle(GraphicHandle), -1, sci_handles, nbRow, nbCol);
+        setMenuParent(sciGetPointerFromHandle(GraphicHandle), (size_t)(-1), sci_handles, nbRow, nbCol);
     }
 
     /* Create return variable */
